+++ esp-sync/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Added `#[inline]` attributes to ensure better codegen (#4627)
 
 ### Fixed